fix:価格登録のエラー解消
